Allow specifying IPs manually on the command line.
There are some unique cases where this is necessary such as when using
network namespaces to isloate the IP assignment on a given host.

For example, one could use a container which uses a bridge IPVLAN
interface to ensure the shared IP is available online 'inside' that
container. When this is the case, you cannot run lelastic (or any other
BGP client) inside the container as it requires it's own IPv6 network
assigned via SLAAC to communicate with the peers.

As you have to run it one the host, and the IPs in question are not
shown on any interface visible on the host, they cannot be
auto-detected.

The simplest solution here is to allow specifying the IPs on the command
line.

A future improvement might be to speak to systemd-machined to obtain a
list of machine and their relevant IPs to advertise (assuming the
container is registered with systemd-machined).